# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Filament DB API",
    "description": "REST API for managing 3D printing filament profiles, spools, nozzles, printers, bed types, locations, print history, sharing, slicer integrations, and OpenPrintTag imports.",
    "version": "1.17.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/api/print-history": {
      "get": {
        "tags": ["Print History"],
        "summary": "List print history",
        "parameters": [
          { "name": "filamentId", "in": "query", "schema": { "type": "string" }, "description": "Filter to jobs using a filament" },
          { "name": "printerId", "in": "query", "schema": { "type": "string" }, "description": "Filter to a printer" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 }, "description": "Maximum rows to return" }
        ],
        "responses": {
          "200": { "description": "Print history entries", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } }
        }
      },
      "post": {
        "tags": ["Print History"],
        "summary": "Record print job",
        "description": "Creates a PrintHistory row, debits referenced spools, and stamps usageHistory entries with the job id for safe undo.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["jobLabel", "usage"],
                "properties": {
                  "jobLabel": { "type": "string", "maxLength": 200 },
                  "printerId": { "type": "string", "nullable": true },
                  "startedAt": { "type": "string", "format": "date-time" },
                  "source": { "type": "string", "enum": ["manual", "prusaslicer", "orcaslicer", "bambu", "other"] },
                  "notes": { "type": "string", "maxLength": 2000 },
                  "usage": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 100,
                    "items": {
                      "type": "object",
                      "required": ["filamentId", "grams"],
                      "properties": {
                        "filamentId": { "type": "string" },
                        "spoolId": { "type": "string" },
                        "grams": { "type": "number", "minimum": 0 }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created print history entry", "content": { "application/json": { "schema": { "type": "object" } } } },
          "400": { "description": "Invalid print history body", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Referenced filament not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
...
    }
  }
}